feat: fix issue with array types in deps
Array types are prefixed with an underscore in pg_type. So when we do
matching inside of load_deps, we were not matching it properly because
it would look for foo and we would show _foo as the dependency. This
would fail the match and then break,

This approach is naive but simply removes the _ if it's an array type.
Not sure if this will cause any other issues.